requirements.txt
streamlit==1.49.1
chromadb==1.0.21
sentence-transformers==5.1.0
google-generativeai==0.8.5
scikit-learn==1.7.2
joblib==1.5.2
pandas==2.3.2
python-dotenv==1.1.1
langchain